+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 - csharp: converting function on index page
+- csharp-mvc: converting function on index page
 - editing by default
 - change reference source
 - filename in editor page title