+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - csharp: converting function on index page
 - csharp-mvc: converting function on index page
+- java: converting function on index page
 - editing by default
 - change reference source
 - filename in editor page title