+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - csharp: converting function on index page
 - csharp-mvc: converting function on index page
 - java: converting function on index page
+- java-spring: converting function on index page
 - editing by default
 - change reference source
 - filename in editor page title