+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - csharp-mvc: converting function on index page
 - java: converting function on index page
 - java-spring: converting function on index page
+- php: converting function on index page
 - editing by default
 - change reference source
 - filename in editor page title