+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - java: converting function on index page
 - java-spring: converting function on index page
 - php: converting function on index page
+- php-laravel: converting function on index page
 - editing by default
 - change reference source
 - filename in editor page title