+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - php: converting function on index page
 - php-laravel: converting function on index page
 - python: converting function on index page
+- ruby: converting function on index page
 - editing by default
 - change reference source
 - filename in editor page title